Started writing the disassembler

+++ Base.hpp
@@ -1,14 +1,17 @@
 #pragma once
 #include <type_traits>
 #include <functional>
+#include <iostream>
 #include <typeinfo>
 #include <memory>
+#include <vector>
+#include <set>
 #include <map>
 
 typedef unsigned char UInt8; typedef char Int8;
-typedef unsigned short UInt16; typedef short Int16;
-typedef unsigned UInt32; typedef int Int32; typedef float  Float32;
-typedef unsigned long UInt64; typedef long int Int64; typedef double Float64;
+typedef unsigned short int UInt16; typedef short int Int16;
+typedef unsigned long int UInt32; typedef long int Int32; typedef float Float32;
+typedef unsigned long long int UInt64; typedef long long int Int64; typedef double Float64;
 
 namespace std {
 	template <bool condition, typename TrueType, TrueType trueValue, typename FalseType, FalseType falseValue>
